--- doc/unreleased/default/back_to_the_future.md
@@ -1,9 +1,0 @@
-# Faker::BackToTheFuture
-
-```ruby
-Faker::BackToTheFuture.character #=> "Marty McFly"
-
-Faker::BackToTheFuture.date #=> "November 5, 1955"
-
-Faker::BackToTheFuture.quote #=> "Roads? Where we're going, we don't need roads."
-```
