+++ lib/faker/movies/back_to_the_future.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module Faker
+  module Movies
+    class BackToTheFuture < Base
+      class << self
+        def character
+          fetch('back_to_the_future.characters')
+        end
+
+        def date
+          fetch('back_to_the_future.dates')
+        end
+
+        def quote
+          fetch('back_to_the_future.quotes')
+        end
+      end
+    end
+  end
+end
