+++ test/deprecate/test_deprecate_back_to_the_future.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
+require_relative '../test_helper'
 
-class TestFakerBackToTheFuture < Test::Unit::TestCase
+class TestDeprecateBackToTheFuture < Test::Unit::TestCase
   def setup
     @tester = Faker::BackToTheFuture
   end
