+++ test/faker/movies/test_faker_back_to_the_future.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require_relative '../../test_helper'
+
+class TestFakerBackToTheFuture < Test::Unit::TestCase
+  def setup
+    @tester = Faker::Movies::BackToTheFuture
+  end
+
+  def test_character
+    assert @tester.character.match(/\w+/)
+  end
+
+  def test_date
+    assert @tester.date.match(/\w+/)
+  end
+
+  def test_quote
+    assert @tester.quote.match(/\w+/)
+  end
+end
